requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioredis==1.3.1
aiorwlock==1.3.0
aiosignal==1.2.0
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
backoff==1.10.0
beautifulsoup4==4.11.1
betfairlightweight==2.17.0
betfairutil==0.4.0
black==22.6.0
bleach==5.0.1
blessed==1.19.1
blessings==1.7
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
chex==0.1.5
click==8.0.4
cloudpickle==2.1.0
colorama==0.4.5
colorful==0.5.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.6
dm-tree==0.1.7
docker-pycreds==0.4.0
entrypoints==0.4
etils==0.8.0
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.16.2
filelock==3.7.1
flatbuffers==22.9.24
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.8.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.29
google-api-core==2.8.2
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
gpustat==1.0.0
grpcio==1.43.0
gym==0.24.1
gym-notices==0.0.7
h11==0.13.0
h5py==3.7.0
higher==0.2.1
hiredis==2.0.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.10.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
iso8601==1.0.2
isoduration==20.11.0
itables==1.3.0
jax==0.3.23
jaxlib==0.3.22
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.18.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kopf==1.35.4
kubernetes==23.6.0
libclang==14.0.6
llvmlite==0.38.1
lz4==4.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==2.0.4
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.13
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.0
nest-asyncio==1.5.5
notebook==6.5.2
notebook_shim==0.2.2
numba==0.55.2
numpy==1.22.2
nvidia-ml-py==11.495.46
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencensus==0.9.0
opencensus-context==0.1.2
opencv-python-headless==3.4.18.65
opentelemetry-api==1.1.0
opentelemetry-exporter-otlp==1.1.0
opentelemetry-exporter-otlp-proto-grpc==1.1.0
opentelemetry-proto==1.1.0
opentelemetry-sdk==1.1.0
opentelemetry-semantic-conventions==0.20b0
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathos==0.2.9
pathspec==0.9.0
pathtools==0.1.2
perfetto==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
pox==0.3.1
ppft==1.7.6.5
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-spy==0.3.12
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2021.3
PyYAML==6.0
pyzmq==25.0.0
ray==2.0.0
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.8
scikit-learn==1.1.3
scipy==1.7.3
seaborn==0.12.0
Send2Trash==1.8.0
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
starlette==0.19.1
tabulate==0.8.9
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado==6.2
traitlets==5.8.1
tune-sklearn==0.4.3
typing_extensions==4.0.1
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.16.0
virtualenv==20.16.5
wandb==0.13.4
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.5.2
yarl==1.7.2
zipp==3.8.0
